+++ src/main/java/com/example/apibank/interfaces/AccountService.java
@@ -2,11 +2,13 @@ package com.example.apibank.interfaces;
 
 import com.example.apibank.dtos.EventDto;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
 
     public abstract Object transferEvent(EventDto event);
 
-    public abstract Float balance(String id);
+    public abstract BigDecimal balance(String id);
 
     public abstract void reset();
